+++ terminal_layout/extensions/input/__init__.py
@@ -1,0 +1,1 @@
+from terminal_layout.extensions.input.input_ex import InputEx