+++ terminal_layout/logger.py
@@ -1,9 +1,0 @@
-import logging
-
-logger = logging.getLogger(__name__)
-logger.setLevel(level=logging.DEBUG)
-# handler = logging.FileHandler("terminal_layout.log")
-# handler.setLevel(logging.DEBUG)
-# formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
-# handler.setFormatter(formatter)
-# logger.addHandler(handler)