+++ terminal_layout/view/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from terminal_layout.view.params import Visibility, Gravity, Width
+from terminal_layout.view.params import Visibility, Gravity, Width, Overflow
 from terminal_layout.view.layout import TableLayout, TableRow
 from terminal_layout.view.text_view import TextView
-